jm/prj2

+++ src/main/java/com/example/demo/service/FutsalPartyService.java
@@ -2,7 +2,10 @@ package com.example.demo.service;
 
 import java.util.*;
 
+import javax.xml.transform.*;
+
 import org.springframework.beans.factory.annotation.*;
+import org.springframework.security.core.*;
 import org.springframework.stereotype.*;
 
 import com.example.demo.domain.*;
@@ -38,24 +41,43 @@ public class FutsalPartyService {
 		return cnt == 1;
 	}
 
-	public Map<String, Object> futsalPartyMember(FutsalPartyMember futsalPartyMember) {
-		Map<String, Object> result = new HashMap<>();
-	
-		futsalPartyMember.setFutsalApplyMember("test");
+	public Map<String, Object> futsalPartyMember(FutsalPartyMember futsalPartyMember,
+			Authentication authentication) {
 		
-		result.put("member", false);
+		// 멤버 객체 생성해서 userId로 받기
+		Member member = futsalPartyMapper.selectByUserId(authentication.getName());
+		FutsalParty party = futsalPartyMapper.selectById(futsalPartyMember.getFutsalPartyId());
+		int applyNum = futsalPartyMapper.countByPartyId(futsalPartyMember.getFutsalPartyId());
 		
-		Integer deleteCnt = futsalPartyMapper.deletePartyMember(futsalPartyMember);
-		
-		if(deleteCnt != 1) {
-			Integer insertCnt = futsalPartyMapper.insertPartyMember(futsalPartyMember);
-			result.put("member", true);		
+		if (party.getMemberNum() > applyNum) {
+			
+			futsalPartyMember.setFutsalApplyMember(member.getNickName());
+			
+			Map<String, Object> result = new HashMap<>();
+			
+			result.put("member", false);
+			
+			Integer deleteCnt = futsalPartyMapper.deletePartyMember(futsalPartyMember);
+			
+			if(deleteCnt != 1) {
+				Integer insertCnt = futsalPartyMapper.insertPartyMember(futsalPartyMember);
+				result.put("member", true);		
+			}
+			
+			Integer count = futsalPartyMapper.countByFutsalPartyId(futsalPartyMember.getFutsalPartyId());
+			result.put("count", count);
+			
+			return result;
+			
+			
+		} else {
+			// 신청 불가능한 경우
+			Map<String, Object> result = new HashMap<>();
+			result.put("member", false);
+			result.put("message", "신청이 불가능합니다.");
+			return result;
 		}
 		
-		Integer count = futsalPartyMapper.countByFutsalPartyId(futsalPartyMember.getFutsalPartyId());
-		result.put("count", count);
-		
-		return result;
 	}
